Move functions in time_utils to ESP32Time
